test_apps/v2/python/37/requirements.txt
azure-functions
numpy==1.17.2
azure-storage-blob==12.1.0
google-auth==1.9.0
